seeded database. Removed listItems model. Updated package.json for scripts. Consolidated seed files. Added properties to Order model.
