+++ models/index.js
@@ -1,7 +1,5 @@
 module.exports = {
   User: require('./User'),
   Product: require('./Product'),
-  Menu: require('./Menu'),
   Order: require('./Order'),
-  ListItems: require('./ListItems')
 };
