+++ package.json
@@ -11,8 +11,7 @@
     "start:prod": "node server.js",
     "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
     "client": "cd client && npm run start",
-    "listItemSeed": "node scripts/listItemSeed.js",
-    "orderSeed": "node scripts/orderSeed.js",
+    "seed": "node scripts/seedDB.js",
     "install": "cd client && npm install",
     "build": "cd client && npm run build",
     "heroku-postbuild": "npm run build"
